pyproject.toml
[project]
name = "wombat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "praw==7.8.1",
    "python-dotenv==1.2.1",
]
